+++ setup.py
@@ -2,6 +2,8 @@ from setuptools import setup
 
 with open('README.md') as f:
     long_description = f.read()
+
+
 setup(
     name="pandera",
     version="0.2.0",
@@ -42,4 +44,3 @@ setup(
         ],
 
 )
-
